fix(bitbucket-lfs): move duplicate code to util, add lfs tests, fix test
